Autumn cleaning part 4 - split test/common.py (#477)
