+++ test/fu/test_div_unit.py
@@ -5,7 +5,7 @@ from coreblocks.fu.div_unit import DivFn, DivComponent
 
 from test.fu.functional_common import ExecFn, FunctionalUnitTestCase
 
-from test.common import signed_to_int, int_to_signed
+from transactron._utils import signed_to_int, int_to_signed
 
 
 @parameterized_class(
